[M] chat list user
In this section, the frontend of the chat list that displays the user's contacts has been developed. This task is divided into three files. Firstly, we have 'ChatUserData' located in 'domain/model', which is used to represent the data related to a user in a chat using a data structure.
The following two files are located in 'presentation/interaction/chats'. The first one, called 'ListChatsFragment', is intended for the backend team to deploy their code. The second, 'UsersAdapter', is used to display the list of users in a list view, making use of the data structure provided by 'ChatUserData'.