[M] chat screen
In this section, the frontend of the chat screen that shows the conversation messages has been developed. This task is divided into three files. Firstly, we have 'ChatMessageUser' located in 'domain/model', which is used to represent the data related to a user in a chat using a data structure. The following two files are located in 'presentation/interaction/chats'. The first, called 'ChatScreenFragment', is intended for the backend team to deploy their code. The second, 'ChatAdapter', is used to display each sent message in a list view, making use of the data structure provided by 'ChatMessageUser'.