Create manual-backend-cd.yml
This pipeline is created directly in the main branch, since it is the default branch and according to the Github Actions documentation, to execute pipelines manually, they must be in the default branch. The reason is to be able to execute the deployment pipelines in the develop branch manually, without having to wait for a commit or merge in the branch.