[M] junitTest profile user
Tests were carried out for the conversions of the numbers of posts, photos, following and follows to "k" representing thousands and "M" representing millions. In these tests carried out, it was observed that the conversion of a number such as 1500 was converted to only 1k when it should be 1.5k, this error was corrected and the function was also corrected.