contract login
Furthermore, we have not defined scenarios in which, for example, only the password or only the user_name is entered separately. This is because these validations are performed within the test itself, so when we add a JSON body with such characteristics, the test will pass as it validates that both fields are mandatory. Therefore, the test is already functioning correctly.

#753