Update main.tf
Enabling all the necessary APIs. Eliminating the "data" section that allows obtaining the information of the "default" private connection, which is currently used in the deployed services. However, for this terraform file I have decided to create an example custom VPC, following the structure that will be used in production.